// src/enums/orderStatus.ts
export enum OrderStatusEnum {
  CREATED = "CREATED",
  ACCEPTED = "ACCEPTED",
  READY_TO_PICKUP = "READY_TO_PICKUP",
  ON_THE_WAY = "ON_THE_WAY",
  DELIVERED = "DELIVERED",
}
export enum OrderStatus {
  CREATED = "Commande passée",
  ACCEPTED = "Commande acceptée",
  READY_TO_PICKUP = "Commande en attente de récupération par le livreur",
  ON_THE_WAY = "En route",
  DELIVERED = "Livrée",
}